+++ plugins/action/sda_port_channels_info.py
@@ -32,7 +32,7 @@ argument_spec.update(
         networkDeviceId=dict(type="str"),
         portChannelName=dict(type="str"),
         connectedDeviceType=dict(type="str"),
-        nativeVlanId=dict(type="float"),
+        nativeVlanId=dict(type="int"),
         offset=dict(type="int"),
         limit=dict(type="int"),
         headers=dict(type="dict"),
@@ -102,6 +102,8 @@ class ActionModule(ActionBase):
             function="get_port_channels_connectivity",
             params=self.get_object(self._task.args),
         )
-        self._result.update(dict(catalystcenter_response=response, dnac_response=response))
+        self._result.update(
+            dict(catalystcenter_response=response, dnac_response=response)
+        )
         self._result.update(catalystcenter.exit_json())
         return self._result
